+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_template_builder.py
@@ -405,13 +405,27 @@ def build_vm_resource(  # pylint: disable=too-many-locals
 def _build_data_disks(profile, data_disk_sizes_gb, image_data_disks,
                       data_caching, storage_sku, attach_data_disks=None):
     lun = 0
+
+    # handle 2 kinds of values
+    # 1 "--data-disk-caching <value>": all disks will be applied
+    # 2 "--data-disk-caching 1=<value> 2=<value>": apply based on lun, the rest will use server side default
+    default_caching, individual_disk_cachings = None, {}
+    if data_caching:
+        if len(data_caching) == 1 and '=' not in data_caching[0]:
+            default_caching = data_caching[0]
+        else:
+            for x in data_caching:
+                temp, caching = x.split('=', 1)
+                temp = int(temp)
+                individual_disk_cachings[temp] = caching
+
     if image_data_disks:
         profile['dataDisks'] = profile.get('dataDisks') or []
         for image_data_disk in image_data_disks or []:
             profile['dataDisks'].append({
                 'lun': image_data_disk.lun,
                 'createOption': "fromImage",
-                'caching': data_caching,
+                'caching': default_caching or individual_disk_cachings.get(image_data_disk.lun),
                 'managedDisk': {'storageAccountType': storage_sku}
             })
             lun = lun + 1
@@ -424,7 +438,7 @@ def _build_data_disks(profile, data_disk_sizes_gb, image_data_disks,
                 'lun': lun,
                 'createOption': "empty",
                 'diskSizeGB': int(size),
-                'caching': data_caching,
+                'caching': default_caching or individual_disk_cachings.get(lun),
                 'managedDisk': {'storageAccountType': storage_sku}
             })
             lun = lun + 1
@@ -436,7 +450,7 @@ def _build_data_disks(profile, data_disk_sizes_gb, image_data_disks,
             disk_entry = {
                 'lun': lun,
                 'createOption': 'attach',
-                'caching': data_caching,
+                'caching': default_caching or individual_disk_cachings.get(lun),
             }
             if is_valid_resource_id(d):
                 disk_entry['managedDisk'] = {'id': d}